updated test_00
